Allow Other modules to open existing database

+++ database.js
@@ -23,36 +23,38 @@ import fs from 'node:fs';
 
 
 const debug = Debug('database');
+let db;
 
 export default function (dbfilename, initfile) {
 
-  let db;
+  if (!db) {
 
-  try {
-    db = new Database(dbfilename);
-  } catch(e) {
-    if (e.code === 'SQLITE_CANTOPEN') {
-      //looks like database didn't exist, so we had better make if from scratch
-      try {
-        debug ('could not open database as it did not exist - so now going to create it');
-        db = new Database(dbfilename, { fileMustExist: false, timeout: 5000 });
-        debug('Opened database - ready to start creating structure');
-        const database = fs.readFileSync(initfile, 'utf8');
-        db.exec(database);
-        const pin = 'T' + ('000000' + (Math.floor(Math.random() * 999999)).toString()).slice(-6); //make a new pin 
-        debug('going to use', pin, 'as our token key');
-        db.prepare(`UPDATE settings SET value = ? WHERE name = 'token_key'`).run(pin);
-        debug('Successfully updated blank database with script')
-      } catch (e) {
-        fs.unlinkSync(dbfilename); //failed to create it. so delete it so we can correct problem and try again.
-        throw new Error(`Encountered ${e.toString()} error when trying to create ${dbfilename} or to initialsize from ${initfile}`)
+    try {
+      db = new Database(dbfilename);
+    } catch(e) {
+      if (e.code === 'SQLITE_CANTOPEN') {
+        //looks like database didn't exist, so we had better make if from scratch
+        try {
+          debug ('could not open database as it did not exist - so now going to create it');
+          db = new Database(dbfilename, { fileMustExist: false, timeout: 5000 });
+          debug('Opened database - ready to start creating structure');
+          const database = fs.readFileSync(initfile, 'utf8');
+          db.exec(database);
+          const pin = 'T' + ('000000' + (Math.floor(Math.random() * 999999)).toString()).slice(-6); //make a new pin 
+          debug('going to use', pin, 'as our token key');
+          db.prepare(`UPDATE settings SET value = ? WHERE name = 'token_key'`).run(pin);
+          debug('Successfully updated blank database with script')
+        } catch (e) {
+          fs.unlinkSync(dbfilename); //failed to create it. so delete it so we can correct problem and try again.
+          throw new Error(`Encountered ${e.toString()} error when trying to create ${dbfilename} or to initialsize from ${initfile}`)
+        }
+      } else {
+        throw new Error(`Encountered ${e.toString()} error when opening database`);
       }
-    } else {
-      throw new Error(`Encountered ${e.toString()} error when opening database`);
     }
-  }
 
-  db.pragma('foreign_keys = ON');
-  process.on('exit', () => db.close());
+    db.pragma('foreign_keys = ON');
+    process.on('exit', () => db.close());
+  }
   return db;
 }
